pyproject.toml
[project]
name = "agent-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain (>=1.0.8,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
